fix id conversion

+++ github/resource_github_codespaces_organization_secret.go
@@ -102,7 +102,7 @@ func resourceGithubCodespacesOrganizationSecretCreateOrUpdate(d *schema.Resource
 		ids := selectedRepositories.(*schema.Set).List()
 
 		for _, id := range ids {
-			selectedRepositoryIDs = append(selectedRepositoryIDs, id.(int64))
+			selectedRepositoryIDs = append(selectedRepositoryIDs, int64(id.(int)))
 		}
 	}
